pyproject.toml
# License: MIT
# Copyright © 2023 Frequenz Energy-as-a-Service GmbH

[build-system]
requires = [
  "setuptools == 68.1.0",
  "setuptools_scm[toml] == 7.1.0",
  "frequenz-repo-config[api] == 0.7.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "frequenz-api-reporting"
description = "Frequenz gRPC API to aggregate component data from microgrids"
readme = "README.md"
license = { text = "MIT" }
keywords = ["frequenz", "python", "api", "grpc", "protobuf", "rpc", "reporting", "cloud", "microgrid"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]
requires-python = ">= 3.11, < 4"
dependencies = [
  "frequenz-api-common >= 0.5.1, < 0.6.0",
  "frequenz-client-common @ git+https://github.com/frequenz-floss/frequenz-client-common-python.git@v0.x.x",
  "googleapis-common-protos >= 1.56.2, < 2",
  "grpcio >= 1.51.1, < 2",
]
dynamic = ["version"]

[[project.authors]]
name = "Frequenz Energy-as-a-Service GmbH"
email = "floss@frequenz.com"

[project.optional-dependencies]
dev-flake8 = [
  "flake8 == 6.1.0",
  "flake8-docstrings == 1.7.0",
  "flake8-pyproject == 1.2.3",  # For reading the flake8 config from pyproject.toml
  "pydoclint == 0.3.2",
  "pydocstyle == 6.3.0",
]
dev-formatting = ["black == 23.9.1", "isort == 5.12.0"]
dev-mkdocs = [
  "black == 23.9.1",
  "Markdown==3.4.4",
  "mike == 2.0.0",
  "mkdocs-gen-files == 0.5.0",
  "mkdocs-literate-nav == 0.6.1",
  "mkdocs-macros-plugin == 1.0.4",
  "mkdocs-material == 9.3.1",
  "mkdocstrings[python] == 0.23.0",
  "frequenz-repo-config[api] == 0.7.0",
]
dev-mypy = [
  "mypy == 1.5.1",
  "grpc-stubs == 1.53.0.2",
  "types-Markdown == 3.4.2.10",
  # For checking the noxfile, docs/ script, and tests
  "frequenz-api-reporting[dev-mkdocs,dev-noxfile,dev-pytest]",
]
dev-noxfile = [
  "nox == 2023.4.22",
  "frequenz-repo-config[api] == 0.7.0",
]
dev-pylint = [
  "pylint >= 2.17.3, < 3",
  # For checking the noxfile, docs/ script, and tests
  "frequenz-api-reporting[dev-mkdocs,dev-noxfile,dev-pytest]",
]
dev-pytest = [
  "pytest == 7.4.2",
  "frequenz-repo-config[extra-lint-examples] == 0.7.0",
]
dev = [
  "frequenz-api-reporting[dev-mkdocs,dev-flake8,dev-formatting,dev-mkdocs,dev-mypy,dev-noxfile,dev-pylint,dev-pytest]",
]

[project.urls]
Documentation = "https://frequenz-floss.github.io/frequenz-api-reporting/"
Changelog = "https://github.com/frequenz-floss/frequenz-api-reporting/releases"
Issues = "https://github.com/frequenz-floss/frequenz-api-reporting/issues"
Repository = "https://github.com/frequenz-floss/frequenz-api-reporting"
Support = "https://github.com/frequenz-floss/frequenz-api-reporting/discussions/categories/support"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["benchmarks", "examples", "src", "tests"]

[tool.flake8]
# We give some flexibility to go over 88, there are cases like long URLs or
# code in documenation that have extra indentation. Black will still take care
# of making everything that can be 88 wide, 88 wide.
max-line-length = 100
extend-ignore = [
  "E203", # Whitespace before ':' (conflicts with black)
  "W503", # Line break before binary operator (conflicts with black)
]
# pydoclint options
style = "google"
check-return-types = false
check-yield-types = false
arg-type-hints-in-docstring = false
arg-type-hints-in-signature = true
allow-init-docstring = true

[tool.pylint.similarities]
ignore-comments = ['yes']
ignore-docstrings = ['yes']
ignore-imports = ['no']
min-similarity-lines = 40

[tool.pylint.messages_control]
disable = [
  "too-few-public-methods",
  # disabled because it conflicts with isort
  "wrong-import-order",
  "ungrouped-imports",
  # pylint's unsubscriptable check is buggy and is not needed because
  # it is a type-check, for which we already have mypy.
  "unsubscriptable-object",
  # Checked by flake8
  "line-too-long",
  "unused-variable",
  "unnecessary-lambda-assignment",
]

[tool.pytest.ini_options]
testpaths = ["pytests"]

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
# This option disables mypy cache, and it is sometimes useful to enable it if
# you are getting weird intermittent error, or error in the CI but not locally
# (or vice versa). In particular errors saying that type: ignore is not
# used but getting the original ignored error when removing the type: ignore.
# See for example: https://github.com/python/mypy/issues/2960
#no_incremental = true
packages = ["frequenz.api.reporting"]
strict = true

[[tool.mypy.overrides]]
module = ["mkdocs_macros.*", "sybil", "sybil.*"]
ignore_missing_imports = true

[tool.setuptools.package-dir]
"" = "py"

[tool.setuptools.package-data]
"*" = ["*.pyi"]

[tool.setuptools_scm]
version_scheme = "post-release"
